double -> float for ScreenLang params (pos, alpha, etc), fix some warnings, perfomance +4%

+++ src/gui/screen/hotspot.h
@@ -8,8 +8,8 @@
 
 class Hotspot: public Child {
 private:
-	double scaleX = 1;
-	double scaleY = 1;
+	float scaleX = 1;
+	float scaleY = 1;
 
 	bool prevMouseOver = false;
