+++ src/utils/math.cpp
@@ -3,17 +3,17 @@
 #include <math.h>
 #include <float.h>
 
-static double *sins;
+static float *sins;
 void Math::init() {
 	const double PI = atan(1) * 4;
 
-	sins = new double[91]();
+	sins = new float[91]();
 	for (int i = 0; i <= 90; ++i) {
-		sins[i] = sin(i * (PI / 180));
+		sins[i] = float(sin(i * (PI / 180)));
 	}
 }
 
-double Math::getSin(int i) {
+float Math::getSin(int i) {
 	if (!i) return 0;
 
 	i = i % 360;
@@ -21,13 +21,13 @@ double Math::getSin(int i) {
 		i += 360;
 	}
 
-	double k = i < 180 ? 1 : -1;
+	float k = i < 180 ? 1 : -1;
 	i = i % 180;
 
 	return sins[i < 90 ? i : 180 - i] * k;
 }
 
-double Math::getCos(int i) {
+float Math::getCos(int i) {
 	if (!i) return 1;
 
 	i = (i - 90) % 360;
@@ -35,7 +35,7 @@ double Math::getCos(int i) {
 		i += 360;
 	}
 
-	double k = i > 180 ? 1 : -1;
+	float k = i > 180 ? 1 : -1;
 	i = i % 180;
 
 	return sins[i < 90 ? i : 180 - i] * k;
